+++ src/Control/Carrier/State/IORef.hs
@@ -10,11 +10,14 @@
 
 {- | A carrier for the 'State' effect. It uses an 'IORef' internally to handle its state, and thus admits a 'MonadUnliftIO' instance. Because the state operations are performed impurely, this carrier will not lose state effects even with nefarious uses of 'liftWith'.
 
-Note that the parameter order in 'runState', 'evalState', and 'execState' is reversed compared to the equivalent functions provided by @transformers@. This is an intentional decision made to enable the composition of effect handlers with '.' without invoking 'flip'.
+Unlike the other carriers for 'State', this carrier's effects will not backtrack when run in conjuction with 'NonDet' effects.
+
+@since 1.1.2.0
 -}
 module Control.Carrier.State.IORef
-( -- * Strict state carrier
+( -- * Impure state carrier
   runState
+, runStateRef
 , evalState
 , execState
 , StateC(..)
@@ -38,7 +41,7 @@ import           Data.IORef
 --
 --   prop> run (runState a (pure b)) === (a, b)
 --
--- @since 1.0.0.0
+-- @since 1.1.2.0
 runState :: MonadIO m => s -> StateC s m a -> m (s, a)
 runState s x = do
   ref <- liftIO $ newIORef s
@@ -47,11 +50,23 @@ runState s x = do
   pure (final, result)
 {-# INLINE[3] runState #-}
 
+-- | Run a 'State' effect starting from the passed 'IORef'.
+--
+--   prop> run (runState a (pure b)) === (a, b)
+--
+-- @since 1.1.2.0
+runStateRef :: MonadIO m => IORef s -> StateC s m a -> m (s, a)
+runStateRef ref x = do
+  result <- runReader ref . runStateC $ x
+  final <- liftIO . readIORef $ ref
+  pure (final, result)
+{-# INLINE[3] runStateRef #-}
+
 -- | Run a 'State' effect, yielding the result value and discarding the final state.
 --
 --   prop> run (evalState a (pure b)) === b
 --
--- @since 1.0.0.0
+-- @since 1.1.2.0
 evalState :: forall s m a . MonadIO m => s -> StateC s m a -> m a
 evalState s x = do
   ref <- liftIO $ newIORef s
@@ -62,12 +77,12 @@ evalState s x = do
 --
 --   prop> run (execState a (pure b)) === a
 --
--- @since 1.0.0.0
+-- @since 1.1.2.0
 execState :: forall s m a . MonadIO m => s -> StateC s m a -> m s
 execState s = fmap fst . runState s
 {-# INLINE[3] execState #-}
 
--- | @since 1.0.0.0
+-- | @since 1.1.2.0
 newtype StateC s m a = StateC { runStateC :: ReaderC (IORef s) m a }
   deriving (Alternative, Applicative, Functor, Monad, Fail.MonadFail, MonadFix, MonadIO, MonadPlus, MonadTrans, MonadUnliftIO)