+++ ChangeLog.md
@@ -1,6 +1,6 @@
 # v1.1.2.0
 
-- Adds `MonadUnliftIO` instances for `ReaderC`, `LiftC`, and `InterpretC`.
+- Adds `MonadUnliftIO` instances for `ReaderC`, `LiftC`, and `InterpretC`. ([#420](https://github.com/fused-effects/fused-effects/pull/420))
 
 # v1.1.1.2