+++ ChangeLog.md
@@ -1,5 +1,12 @@
 - Adds `MonadUnliftIO` instances for `ReaderC`, `LiftC`, and `InterpretC`. ([#420](https://github.com/fused-effects/fused-effects/pull/420))
 
+- Adds `Accum` ([#391](https://github.com/fused-effects/fused-effects/pull/391)) (by @turion)
+  - Adds an `Accum` effect
+  - Adds a church-encoded `Control.Carrier.Accum.Church` carrier
+  - Adds a strict `Control.Carrier.Accum.Church` carrier
+  - Defines `Algebra` instances for the two mentioned carriers,
+    and for `Control.Monad.Trans.Accum` from `transformers`
+
 # v1.1.1.2
 
 - Adds support for `ghc` 9.2.1 and `base` 4.16.